pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-ai"
version = "0.1.0"
description = "Nexus WhatsApp-first AI assistant core"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "platformdirs>=4.3",
  "litellm>=1.59.0",
  "websockets>=12.0",
  "apscheduler>=3.10",
  "requests>=2.32",
  "markdownify>=0.13",
  "python-dateutil>=2.9",
  "textual>=0.71.0",
  "google-api-python-client>=2.170.0",
  "google-auth>=2.40.0",
  "google-auth-oauthlib>=1.2.2",
]

[project.scripts]
nexus = "nexus.cli_app:main"

[tool.setuptools.packages.find]
include = ["nexus*"]

[tool.setuptools.package-data]
nexus = [
  "prompts/*.md",
  "skills/*/SKILL.md",
  "bridge_runtime/.env.example",
  "bridge_runtime/package.json",
  "bridge_runtime/tsconfig.json",
  "bridge_runtime/src/*.ts",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
